124356723.vbs
Being Bored? Wanna prank your friend and not actually ruin their computer This is a fake virus file  made from VBScript it is full of chaos WARNING! Do not use this file to break someone else friendship this app is only made to play few pranks with your friend and say This was just a prank! So do NOT use this file for illegal and ruining stuff I hope you are able to download this file or else it will be a problem for others and this app is only made for Windows not Apple or Linux. Luckily a billions of people use Windows. If you have Apple or Linux then I am very sorry you cannot install this file